+++ app/models/support_request.rb
@@ -1,0 +1,6 @@
+class SupportRequest < ApplicationRecord
+
+  belongs_to :user
+
+  validates :message, presence: true
+end