+++ app/models/user.rb
@@ -1,4 +1,5 @@
 # A User may have many {Journey}s.
 class User < ApplicationRecord
   has_many :journeys
+  has_many :support_requests
 end