+++ spec/factories/support_requests.rb
@@ -1,0 +1,15 @@
+FactoryBot.define do
+  factory :support_request do
+    message { "Support request message from a School Buying Professional" }
+
+    association :user, factory: :user
+
+    # trait :with_specification_ids do
+    #   # association { create(:journey) }
+    # end
+
+    # trait :with_category_ids do
+    #   # association { create(:create) }
+    # end
+  end
+end