+++ spec/models/user_spec.rb
@@ -5,4 +5,9 @@ RSpec.describe User, type: :model do
     user = build(:user, dfe_sign_in_uid: "a96ee522-2de8-4eb1-9b17-9619a310149b")
     expect(user.dfe_sign_in_uid).to eql("a96ee522-2de8-4eb1-9b17-9619a310149b")
   end
+
+  it "has many support requests" do
+    association = described_class.reflect_on_association(:support_requests)
+    expect(association.macro).to eq(:has_many)
+  end
 end